Extract git module and remove init function
